feat: Add question ID to response from Swagger API
When a question is asked and an ID is generated, that ID is not
communicated to the user. One of the requests that has come up is for
that ID to be returned in the response instead of just a "success"
message. While making this change, I also noticed that there was an
instance where we could fail to send & store the question but still
report a success state. With a little refactoring, this problem has also
been fixed and the code made a little cleaner.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>