fix: Dump QAExchangeRecord on /send-question submission
After feedback from @reflectivedevelopment, it was recommended to just
dump the exchange record (following other conventions in ACA-Py) rather
than returning a customized response. Doing so returns the question_id
(as was the original request) along with it's state. The output format
is similar to the get-questions list call since the get-questions call
also returns the QAExchangeRecord.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>