fix: Fix errors loading the plugin in 0.7.5+
Some changes made in ACA-Py 0.7.5 and beyond have failed the QA plugin
to silently fail to load correctly. Part of the fault is the QA plugin's
own fault (incorrect version in definition) and part is ACA-Py's (now
sending events that the QA plugin isn't supposed to handle). Both of
these deficiencies should now be fixed.

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>